functions/requirements.txt
firebase-admin==6.5.0
functions-framework==3.5.0
google-generativeai==0.7.2
langchain==0.3.7
httpx==0.27.2
pydantic==2.9.2

